A hack to fix view scaling
At least on my machine, trying to preview or render at ≠1x scaling
didn't work as intended: instead of scaling the view down or up to fill
the new size, it simply trimmed or padded the view. That is, I suspect,
because scaling on the root node is disregarded. Hence this hack that
wraps the view in yet another view, thus making the scaling take effect.

Might to revert this in the future if the upstream behaviour changes.